+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 ### Release Notes
+
 All the updates to this exporter are documented in this file.
 
-## 1.0.0 
+## 1.0.0
+
 Initial exporter version with colors to test import into pando-web
 
 ### 🚀 New
 
-## 2.0.0 
+## 2.0.0
+
 Token exporter without panes or assets, to follow. Subject to changes.
+
+## 3.0.0
+
+Updates to naming in outputs, removed elevation tokens.
+
+## 4.0.0
+
+removes deprecated tokens, converts fontsize from rem to px, adds elevation tokens back, adds other dimension tokens.